+++ xbird-starter-dubbo/src/main/java/com/github/zhycn/dubbo/autoconfigure/DubboAutoConfiguration.java
@@ -14,25 +14,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.zhycn.retrofit2.annotations;
+package com.github.zhycn.dubbo.autoconfigure;
 
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
+import org.springframework.context.annotation.Configuration;
 
-import org.springframework.context.annotation.Import;
+/**
+ * Dubbo Auto-configuration
+ * 
+ * @author zhycn
+ * @since 2.0.0 2018-04-12
+ */
+@Configuration
+public class DubboAutoConfiguration {
 
-import com.github.zhycn.retrofit2.boot.RetrofitServiceFactoryBeanRegistrar;
-
-@Target(ElementType.TYPE)
-@Retention(RetentionPolicy.RUNTIME)
-@Documented
-@Import(RetrofitServiceFactoryBeanRegistrar.class)
-public @interface RetrofitClientScan {
-
-  String[] value() default {};
-
-  Class<?>[] basePackageClasses() default {};
 }